requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
googletrans
google-cloud-translate
pathlib
textwrap
pillow
pyttsx3
speech_recognition
langdetect